% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{upload_archive}
\alias{delete_archive}
\alias{upload_archive}
\title{Archives}
\usage{
upload_archive(vault, contents, ...)

delete_archive(vault, archive, ...)
}
\arguments{
\item{vault}{A character string specifying the name of an AWS Glacier vault (max. 255 characters).}

\item{contents}{}

\item{\dots}{Additional arguments passed to \code{\link{glacierHTTP}}.}
}
\description{
Upload or delete Archives
}

