% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{get_glacier_job}
\alias{get_glacier_job}
\alias{get_glacier_job_output}
\title{Get job}
\usage{
get_glacier_job(vault, job, ...)

get_glacier_job_output(vault, job, ...)
}
\arguments{
\item{vault}{A character string specifying the name of an AWS Glacier vault (max. 255 characters).}

\item{job}{\dots}

\item{\dots}{Additional arguments passed to \code{\link{glacierHTTP}}.}
}
\value{
A list.
}
\description{
Get job or job output
}
