% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{glacierHTTP}
\alias{glacierHTTP}
\title{glacierHTTP}
\usage{
glacierHTTP(verb, action, query = list(), headers = list(), body = "",
  region = Sys.getenv("AWS_DEFAULT_REGION", "us-east-1"),
  key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"), ...)
}
\arguments{
\item{verb}{A character string specifying an HTTP method.}

\item{action}{A character string specifying an API \dQuote{action}.}

\item{query}{A list of HTTP URL query parameters.}

\item{headers}{A list of HTTP headers.}

\item{body}{The body of the request.}

\item{region}{A character string containing the AWS region.
If missing, defaults to \dQuote{us-east-1}.}

\item{key}{A character string containing an AWS Access Key ID. 
If missing, defaults to value stored in environment variable \dQuote{AWS_ACCESS_KEY_ID}.}

\item{secret}{A character string containing an AWS Secret Access Key. 
If missing, defaults to value stored in environment variable \dQuote{AWS_SECRET_ACCESS_KEY}.}

\item{\dots}{Additional arguments passed to \code{\link[httr]{GET}}, etc.}
}
\value{
A list.
}
\description{
AWS Glacier API Requests
}
\details{
This is the workhorse function for executing Glacier API requests. It is typically not necessary to call this function directly.
}

