% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{initiate_job}
\alias{initiate_job}
\title{Initiate job}
\usage{
initiate_job(description, type, archive, format, limit, marker, bytes, start,
  end, topic, ...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{glacierHTTP}}.}
}
\description{
Initiate a Glacier job
}

