% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{list_jobs}
\alias{list_jobs}
\title{List jobs}
\usage{
list_jobs(vault, n, completed, marker, status, ...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{glacierHTTP}}.}
}
\description{
List jobs
}

