% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{get_retrieval_policy}
\alias{get_retrieval_policy}
\alias{set_retrieval_policy}
\alias{get_glacier_capacity}
\title{Data Retrieval policies}
\usage{
get_retrieval_policy(...)

set_retrieval_policy(strategy, bytes, ...)

get_glacier_capacity(...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{glacierHTTP}}.}

\item{strategy}{A character string specifying the type of data retrieval policy to set. Must be one of: \dQuote{BytesPerHour}, \dQuote{FreeTier}, or \dQuote{None}.}

\item{bytes}{If \code{strategy = "BytesPerHour"}, an integer specifying the maximum number of bytes that can be retrieved per hour (between 1 and \eqn{2^63-1}).}
}
\value{
A list.
}
\description{
Get and set data retrieval policies
}
