% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{create_vault}
\alias{create_vault}
\alias{delete_vault}
\alias{describe_vault}
\title{Vaults}
\usage{
create_vault(vault, ...)

delete_vault(vault, ...)

describe_vault(vault, ...)
}
\arguments{
\item{vault}{A character string specifying the name of an AWS Glacier vault (max. 255 characters).}

\item{\dots}{Additional arguments passed to \code{\link{glacierHTTP}}.}
}
\description{
Create, describe, or delete vaults
}
