\name{aws.glacier-package}
\alias{aws.glacier-package}
\alias{aws.glacier}
\docType{package}
\title{AWS Glacier Client Package}
\description{A simple client package for the Amazon Web Services (AWS) Glacier  API.}
\details{}
\author{Thomas J. Leeper <thosjleeper@gmail.com>}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}
