% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.glacier-package.R
\docType{package}
\name{aws.glacier-package}
\alias{aws.glacier-package}
\alias{aws.glacier}
\title{aws.glacier}
\description{
AWS Glacier Client Package
}
\details{
A simple client package for the Amazon Web Services (AWS) Glacier  API.
}
\references{
\url{http://docs.aws.amazon.com/amazonglacier/latest/dev/introduction.html}
}
\seealso{
\code{\link{create_vault}}, \code{\link{initiate_glacier_job}}, \code{\link{upload_archive}}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
