% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.glacier-package.R
\docType{package}
\name{aws.glacier-package}
\alias{aws.glacier}
\alias{aws.glacier-package}
\title{aws.glacier}
\description{
AWS Glacier Client Package
}
\details{
A simple client package for the Amazon Web Services (AWS) Glacier  API.
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}

