% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{glacierHTTP}
\alias{glacierHTTP}
\title{glacierHTTP}
\usage{
glacierHTTP(verb, action, query = list(), headers = list(), body = "",
  verbose = getOption("verbose", FALSE),
  region = Sys.getenv("AWS_DEFAULT_REGION", "us-east-1"), key = NULL,
  secret = NULL, session_token = NULL, ...)
}
\arguments{
\item{verb}{A character string specifying an HTTP method.}

\item{action}{A character string specifying an API \dQuote{action}.}

\item{query}{A list of HTTP URL query parameters.}

\item{headers}{A list of HTTP headers.}

\item{body}{The body of the request.}

\item{verbose}{A logical indicating whether to be verbose. Default is given by \code{options("verbose")}.}

\item{region}{A character string specifying an AWS region. See \code{\link[aws.signature]{locate_credentials}}.}

\item{key}{A character string specifying an AWS Access Key. See \code{\link[aws.signature]{locate_credentials}}.}

\item{secret}{A character string specifying an AWS Secret Key. See \code{\link[aws.signature]{locate_credentials}}.}

\item{session_token}{Optionally, a character string specifying an AWS temporary Session Token to use in signing a request. See \code{\link[aws.signature]{locate_credentials}}.}

\item{\dots}{Additional arguments passed to \code{\link[httr]{GET}}, etc.}
}
\value{
A list.
}
\description{
AWS Glacier API Requests
}
\details{
This is the workhorse function for executing Glacier API requests. It is typically not necessary to call this function directly.
}
