% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{get_vault_notification}
\alias{delete_vault_notification}
\alias{get_vault_notification}
\alias{set_vault_notification}
\title{Vault notifications}
\usage{
get_vault_notification(vault, ...)

set_vault_notification(vault, events, sns_topic, ...)

delete_vault_notification(vault, ...)
}
\arguments{
\item{vault}{A character string specifying the name of an AWS Glacier vault (max. 255 characters).}

\item{sns_topic}{A character string specifying an SNS topic name or ARN.}

\item{\dots}{Additional arguments passed to \code{\link{glacierHTTP}}.}

\item{\dots}{Additional arguments passed to \code{\link{glacierHTTP}}.}
}
\description{
Get, set, and delete vault notifications
}

