% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{get_retrieval_policy}
\alias{get_retrieval_policy}
\alias{set_retrieval_policy}
\title{Retrieval policies}
\usage{
get_retrieval_policy(...)

set_retrieval_policy(strategy, bytes, ...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{glacierHTTP}}.}
}
\value{
A list.
}
\description{
Get and set retrieval policies
}

